add a prescribed atmosphere in its own module
